Website/requirements.txt
fastapi==0.70.0
numpy==1.19.5
Pillow==8.4.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.10.0+cpu
torchvision==0.11.1+cpu
uvicorn==0.15.0
aiofiles==0.6.0
python-multipart==0.0.5
jinja2==2.11.2
gunicorn==20.1.0